Request placement
This commit implements request placement via the
koha-plugin-api-reprintsdesk API. Handling for failures is included,
including logging for success and failure.

Also included is a processor for sending the link to the borrower
for a completed article request. This is based on the RapidILL
equivalent and will need further work to be compatible with Reprints
Desk