Checkavailability processor
This processor is designed to be running on a scheduled cron, like the other processors in this repo.
It'll grab 'NEW' requests in the ReprintsDesk backend and send their respective DOI/PMID
to ReprintsDesk's ArticleShelf_CheckAvailability webservice.

It'll parse those returned and set them to 'READY' if immediatelly available.

Also in this commit is:

Added Order_GetPriceEstimate2 API call to CheckAvailability processor

This Order_GetPriceEstimate2 call will be performed for 'NEW' requests not returned by ArticleShelf_CheckAvailability.
These requests are not immediately available in ReprintsDesk and their price needs to be checked.
If the request is missing metadata required for price check (ISSN/ISBN or Year), it'll be put in 'ERROR' status with
a staff note informing as such.
If the request has all the required metadata and a price check is successful, it'll update the price and set its
status to 'STANDBY'.