New MARK_NEW backend action
This action will allow the staff member to mark a request as 'NEW' if the request has the 'ERROR' status.
This is so that if the request has errored from ReprintsDesk, staff can fix the error and then place the request
back to 'NEW' so that a new availability+price check is performed again.