Year mapping and bugfix
Previously, find_core_to_reprints_desk was returning date instead of year,
this caused ILL to not know the year of the request since 'year' is what is used.

Fixed a bug in find_core_to_reprints_desk, its supposed to return prop, not field.
This was likely caused by a copy/paste of find_reprints_desk_property which returns field,
as it should.