Fix PlaceOrders
This is doing 2 things:
- Hiding the 'Mark request READY' button if a request already has an orderID.
This is to prevent staff members from placing it back to 'READY' if it has already
been placed with ReprintsDesk. We keep the 'Mark request READY' button otherwise
because it may 'ERROR' without a orderID and staff can still mark it 'READY' if that
is the case.

- PlaceOrders now only picks up 'READY' requests with orderID = NULL
This is a fail-safe redundancy to the above, it should never happen that a 'READY' request
has a orderID, but we're still making sure that this cronjob is idempotent and
won't attempt to re-place an order with ReprintsDesk that already has been placed in the past